Support for flexible side-chain is added and tested
The flexible part is read-in and merged with the ligand, same as how the
Autodock works.
setAbout works only on the ligand, not the flex part.
twsit (torsions applied to rotatable bonds) is applied to all torsions,
including ligand and the flexible part.
transform (translation and rotation) is applied to the ligand part only.